pyproject.toml
[tool.poetry]
name = "django-zenslackchat"
version = "1.0.0"
description = ""
authors = []
license = "MIT"
readme = "README.rst"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q -p no:warnings --reuse-db"
DJANGO_SETTINGS_MODULE = "webapp.settings"

[tool.poetry.dependencies]
python = "^3.11"
pdpyras = ">=5.2.0"
zenpy = ">=2.0.50"
emoji = ">=2.12.1"
bs4 = ">=0.0.2"
markdown = ">=3.7"
slackclient = ">=2.9.4"
slackeventsapi = ">=3.0.2"
python-dateutil = ">=2.9.0.post0"
python-decouple = ">=3.8"
psycopg2-binary = ">=2.9.9"
requests = ">=2.32.3"
sentry-sdk = ">=2.13.0"
celery = {extras = ["redis"], version = ">=5.4.0"}
dj-database-url = ">=2.2.0"
django-log-formatter-ecs = ">=0.0.5"
djangorestframework = ">=3.15.2"
django-environ = ">=0.11.2"
django-staff-sso-client = ">=4.3.0"
urllib3 = ">=2.2.2"
whitenoise = ">=6.7.0"
waitress = ">=3.0.1"
async-timeout = ">=4.0.3"
oauthlib = ">=3.2.2"
certifi = ">=2024.7.4"
aiohttp = "3.10.11"
typing-extensions = ">=4.12.2"
django = ">=4.2.18"
werkzeug = ">=3.0.6"
jinja2 = ">=3.1.5"
dbt-copilot-python = ">=0.2.2"

[tool.poetry.group.dev.dependencies]
black = ">=24.8.0"
isort = ">=5.13.2"
pytest = ">=8.3.2"
pytest-cov = ">=5.0.0"
pytest-django = ">=4.8.0"
flake8 = ">=7.1.1"
ipython = ">=8.26.0"
dbt-copilot-python = ">=0.2.2"

[tool.djlint]
profile="django"
line_break_after_multiline_tag=true
use_gitignore = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
